Drop UPDATE..SET support from Spark
