+++ src/languages/spark/spark.formatter.ts
@@ -23,8 +23,6 @@ const reservedCommands = expandPhrases([
   // Data manipulation
   'INSERT [INTO | OVERWRITE] [TABLE]',
   'VALUES',
-  'UPDATE', // TODO: not supported
-  'SET', // TODO: not supported
   // DDL
   'ALTER COLUMN',
   'ALTER DATABASE',
