+++ test/features/update.ts
@@ -1,0 +1,40 @@
+import dedent from 'dedent-js';
+
+import { FormatFn } from 'src/sqlFormatter';
+
+export default function supportsUpdate(format: FormatFn) {
+  it('formats simple UPDATE statement', () => {
+    const result = format(
+      "UPDATE Customers SET ContactName='Alfred Schmidt', City='Hamburg' WHERE CustomerName='Alfreds Futterkiste';"
+    );
+    expect(result).toBe(dedent`
+      UPDATE
+        Customers
+      SET
+        ContactName = 'Alfred Schmidt',
+        City = 'Hamburg'
+      WHERE
+        CustomerName = 'Alfreds Futterkiste';
+    `);
+  });
+
+  it('formats UPDATE statement with AS part', () => {
+    const result = format(
+      'UPDATE customers SET total_orders = order_summary.total  FROM ( SELECT * FROM bank) AS order_summary',
+      { aliasAs: 'always' }
+    );
+    expect(result).toBe(dedent`
+      UPDATE
+        customers
+      SET
+        total_orders = order_summary.total
+      FROM
+        (
+          SELECT
+            *
+          FROM
+            bank
+        ) AS order_summary
+    `);
+  });
+}
